% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_triplets_per_locus(2).R
\name{count_triplets_per_locus}
\alias{count_triplets_per_locus}
\title{FUNCTION TO COUNT TRIPLETS FOR EACH LOCUS}
\usage{
count_triplets_per_locus(sequence_df)
}
\arguments{
\item{sequence_df}{A data frame containing two columns:
\itemize{
  \item locus: The genomic loci as input (e.g. "chr1:1000-2000").
  \item sequence: The DNA sequence corresponding to each locus.
}}
}
\value{
A list where each element corresponds to a locus and contains a table of triplet counts for that locus.
}
\description{
This function takes a data frame of DNA sequences from genomic loci and counts the triplets (3-mers) present in each sequence.
In addition, if the central base of a triplet is either A or G, its reverse complement is considered.
}
\examples{
loci <- c("chr1:1000000-1000100", "chr2:2000000-2000100")
library(BSgenome.Hsapiens.UCSC.hg38)
sequence_df <- extract_sequences(loci, BSgenome.Hsapiens.UCSC.hg38)
print(sequence_df)

triplet_counts <- count_triplets_per_locus(sequence_df)
print(triplet_counts)
}
