% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_signatures(5).R
\name{decompose_signatures}
\alias{decompose_signatures}
\title{FUNCTION TO DECOMPOSE MUTATIONAL SIGNATURES}
\usage{
decompose_signatures(cosmic_counts, cosmic_path)
}
\arguments{
\item{cosmic_counts}{Either a list of named numeric vectors (COSMIC counts per locus)
or a single named numeric vector (aggregated COSMIC counts).}

\item{cosmic_path}{The path to the COSMIC signatures file in Alexandrov format
(e.g. COSMIC v3.4 SBS GRCh38).}
}
\value{
A list of decomposed mutational signatures: if the input is a list of COSMIC counts,
a list of results per locus, if input is aggregated COSMIC counts, a single result.
}
\description{
This function performs decomposition of mutational signatures based on COSMIC counts.
It takes as input either a list of COSMIC counts per locus or aggregated COSMIC counts,
and determines the contribution of each mutational signature.
}
\examples{
loci <- c("chr1:1000000-1000100", "chr2:2000000-2000100")
library(BSgenome.Hsapiens.UCSC.hg38)
sequence_df <- extract_sequences(loci, BSgenome.Hsapiens.UCSC.hg38)
print(sequence_df)
triplet_counts <- count_triplets_per_locus(sequence_df)
print(triplet_counts)
global_triplet_counts <- count_triplets_aggregated(triplet_counts)
print(global_triplet_counts)
cosmic_path <- system.file("extdata", "COSMIC_v3.4_SBS_GRCh38 2.txt", package = "MutSignature")
cosmic_counts_per_locus <- map_to_cosmic(triplet_counts, cosmic_path)
print(cosmic_counts_per_locus)
cosmic_counts_aggregated <- map_to_cosmic(global_triplet_counts, cosmic_path)
print(cosmic_counts_aggregated)

# Decompose per locus
result_per_locus <- decompose_signatures(cosmic_counts_per_locus, cosmic_path)
print(result_per_locus)

# Decompose aggregated counts
result_aggregated <- decompose_signatures(cosmic_counts_aggregated, cosmic_path)
print(result_aggregated)
}
