% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sequences(1).R
\name{extract_sequences}
\alias{extract_sequences}
\title{EXTRACT DNA SEQUENCES FROM GENOMIC LOCI}
\usage{
extract_sequences(loci, genome)
}
\arguments{
\item{loci}{A character vector where each element represents a genomic locus in the format "chr:start-end".
For example, "chr1:1000-2000".}

\item{genome}{A reference genome object of class BSgenome (e.g. BSgenome.Hsapiens.UCSC.hg38).
#' @return A data frame with the following columns:
\itemize{
  \item locus: The genomic loci given as input.
  \item sequence: The corresponding DNA sequences.
}}
}
\description{
This function extracts DNA sequences from a set of genomic loci provided as input,
using a reference genome object of class BSgenome.
}
\examples{
loci <- c("chr1:1000000-1000100", "chr2:2000000-2000100")
library(BSgenome.Hsapiens.UCSC.hg38)
sequence_df <- extract_sequences(loci, BSgenome.Hsapiens.UCSC.hg38)
print(sequence_df)
}
