% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mutations(3).R
\name{generate_mutations}
\alias{generate_mutations}
\title{FUNCTION TO GENERATE POSSIBLE MUTATIONS FOR EACH TRIPLET}
\usage{
generate_mutations(triplet)
}
\arguments{
\item{triplet}{A character string representing a DNA triplet (e.g., "ACC").}
}
\value{
A vector of all possible mutations in the COSMIC format.
}
\description{
This function generates all possible mutations for a given triplet in the COSMIC format,
by replacing the central base of the triplet with all other bases except itself.
This function
}
\keyword{internal}
