% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_cosmic(4).R
\name{map_single_triplet_set}
\alias{map_single_triplet_set}
\title{MAP A SINGLE SET OF TRIPLET COUNTS TO THE COSMIC FORMAT}
\usage{
map_single_triplet_set(single_triplet_counts, cosmic_matrix)
}
\arguments{
\item{single_triplet_counts}{A numeric vector of triplet counts.}

\item{cosmic_matrix}{A COSMIC mutational signatures matrix loaded as a data frame.}
}
\value{
A numeric vector of COSMIC counts corresponding to the mutations.
}
\description{
This function maps a single set of triplet counts to the COSMIC mutational signatures format.
}
\keyword{internal}
