% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_cosmic(4).R
\name{map_to_cosmic}
\alias{map_to_cosmic}
\title{FUNCTION TO MAP TRIPLET COUNTS TO THE COSMIC FORMAT}
\usage{
map_to_cosmic(triplet_counts, cosmic_path)
}
\arguments{
\item{triplet_counts}{A list of triplet counts per locus (from count_triplets_per_locus),
or a named numeric vector of aggregated triplet counts (from count_triplets_aggregated).}

\item{cosmic_path}{The path to the COSMIC signatures file in Alexandrov format
(e.g. COSMIC v3.4 SBS GRCh38).}
}
\value{
A list of COSMIC counts for each locus if the input is a list of triplet counts,
or a single vector of aggregated COSMIC counts if the input is a numeric vector.
}
\description{
This function maps triplet counts, either per locus or aggregated, to the COSMIC mutational signatures format.
}
\examples{
loci <- c("chr1:1000000-1000100", "chr2:2000000-2000100")
library(BSgenome.Hsapiens.UCSC.hg38)
sequence_df <- extract_sequences(loci, BSgenome.Hsapiens.UCSC.hg38)
print(sequence_df)
triplet_counts <- count_triplets_per_locus(sequence_df)
print(triplet_counts)
global_triplet_counts <- count_triplets_aggregated(triplet_counts)
print(global_triplet_counts)

cosmic_path <- system.file("extdata", "COSMIC_v3.4_SBS_GRCh38 2.txt", package = "MutSignature")

cosmic_counts_per_locus <- map_to_cosmic(triplet_counts, cosmic_path)
print(cosmic_counts_per_locus)
cosmic_counts_aggregated <- map_to_cosmic(global_triplet_counts, cosmic_path)
print(cosmic_counts_aggregated)
}
